Tidied up the verbose option in various producer functions.

+++ user/eudet/hardware/include/AidaTluPowerModule.hh
@@ -35,11 +35,11 @@ public:
   void led_allRed();
   void led_allOff();
   void led_allWhite();
-  void initI2Cslaves(bool intRef, int verbose);
+  void initI2Cslaves(bool intRef, uint8_t verbose);
   uint32_t _set_bit(uint32_t v, int index, bool x);
-  void setIndicatorRGB(int indicator, const std::array<int, 3>& RGB, int verbose);
+  void setIndicatorRGB(int indicator, const std::array<int, 3>& RGB, uint8_t verbose);
   void setI2CPar( i2cCore  *mycore , char DACaddr, char Exp1Add, char Exp2Add, char IdAdd);
-  void setVchannel(int channel, float voltage, int verbose);
+  void setVchannel(int channel, float voltage, uint8_t verbose);
   void testme();
   void testLED();
   //void SetIntRef(bool intRef, bool verbose);
